+++ src/main.py
@@ -1,4 +1,4 @@
-from src.data import test_widget_info
+from src.widget_data import test_widget_info
 from src.widget import get_hidden_info
